[ALPS02840045] mtk_cpuidle: fix coding style
[Detail]
Change coding style of saving/restoring debug registers

MTK-Commit-Id: d99b18f5aaa0c58f05e3444552f435cdc22610c1

CR-Id: ALPS02840045
Feature: Power Management
(cherry picked from commit 7b4510c52765a874729745a00dc6ba5907fcdded)

+++ drivers/misc/mediatek/watchpoint/mtk_dbg.c
@@ -119,7 +119,7 @@ void print_dbgregs(int cpuid)
 EXPORT_SYMBOL(print_dbgregs);
 #endif
 #if 1
-unsigned int *mt_save_dbg_regs(unsigned int *p, unsigned int cpuid)
+unsigned long *mt_save_dbg_regs(unsigned long *p, unsigned int cpuid)
 {
 	struct wp_trace_context_t *wp_context;
 
@@ -216,7 +216,7 @@ unsigned int *mt_save_dbg_regs(unsigned int *p, unsigned int cpuid)
 	return p;
 }
 
-void mt_restore_dbg_regs(unsigned int *p, unsigned int cpuid)
+void mt_restore_dbg_regs(unsigned long *p, unsigned int cpuid)
 {
 	unsigned long dscr;
 	struct wp_trace_context_t *wp_context;
