+++ drivers/misc/mediatek/watchpoint/mtk_dbg.h
@@ -19,8 +19,8 @@
 
 /*#define DBG_REG_DUMP */
 
-unsigned int *mt_save_dbg_regs(unsigned int *p, unsigned int cpuid);
-void mt_restore_dbg_regs(unsigned int *p, unsigned int cpuid);
+unsigned long *mt_save_dbg_regs(unsigned long *p, unsigned int cpuid);
+void mt_restore_dbg_regs(unsigned long *p, unsigned int cpuid);
 void mt_copy_dbg_regs(int to, int from);
 #ifdef DBG_REG_DUMP
 void dump_dbgregs(int cpuid);
